+++ CHANGELOG.md
@@ -9,3 +9,4 @@
 + Add `ensure_length_of` and `ensure_size_of` aliases to `validate_length_of`.
 + Add `with_minimum` option to `ValidateLengthMatcher`.
 + Add `with_min` and `is_at_least` aliases to `ValidateLengthMatcher#with_minimum`.
++ Add `with_maximum` option to `ValidateLengthMatcher`.